requirements-dev.txt
-r requirements.txt
ruff
black
isort
pylint
pylint-django
mypy
django-stubs[compatible-mypy]
